pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "job-ace"
version = "0.1.0"
description = "Local-first job application assistant"
requires-python = ">=3.11"
authors = [{name = "Job Ace"}]
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "sqlalchemy>=2.0.30",
    "pydantic>=2.7.1",
    "pydantic-settings>=2.0.0",
    "aiosqlite>=0.20.0",
    "httpx>=0.27.0",
    "typer>=0.12.3",
    "structlog>=24.1.0",
    "python-multipart>=0.0.9",
    "playwright>=1.43.0",
    "python-slugify>=8.0.4",
    "jinja2>=3.1.4",
    "orjson>=3.10.3",
    "beautifulsoup4>=4.12.3",
    "pyyaml>=6.0.1"
]

[project.scripts]
job-ace = "cli.app:run"

[project.optional-dependencies]
dev = [
    "pytest>=8.2.2",
    "pytest-asyncio>=0.23.6",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.5"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*", "cli*"]

[tool.ruff]
line-length = 100
select = ["E", "F", "B", "I"]
ignore = ["B008"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q"
pythonpath = ["backend", "cli"]
